requirements.txt
python-telegram-bot==20.7
telethon
flask
waitress